+++ src/components/Modal/Modal.js
@@ -1,117 +1,41 @@
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
-import styled, { css } from 'styled-components';
-import BaseModal from 'react-modal2';
+import styled from 'styled-components';
+import * as focusScope from 'a11y-focus-scope';
+import * as focusStore from 'a11y-focus-store';
 import { DESKTOP } from '../../constants/Breakpoints';
-import { hexToRgba } from '../../helpers/colors';
 
-const PADDINGS = {
-  small: '16px 24px',
-  normal: '24px 48px',
-  large: '24px 48px',
-};
-
-const WIDTHS = {
-  small: 'width: 300px',
-  normal: 'width: 600px',
-  large: 'width: 900px',
-  auto: `
-    width: auto;
-    max-width: 900px;
-  `,
-};
-
-const HeaderStyled = styled.header.attrs({ tabIndex: 0 })`
-  outline: none;
-`;
-
-const FooterStyled = styled.footer``;
-
-const BodyStyled = styled.div``;
-
-export const ModalStyles = css`
-  background-color: ${props => props.theme.white};
-  box-shadow: 0 0 8px 0 rgba(120, 130, 139, 0.5);
-  width: 100%;
-  max-width: 100%;
-  @media ${DESKTOP} {
-    ${props => WIDTHS[props.size]};
-  }
-
-  ${HeaderStyled} {
-    border-bottom: solid 1px ${props => hexToRgba(props.theme.stoneGrey, 0.5)};
-    padding: ${props => PADDINGS[props.size] || PADDINGS.normal};
-    font-size: 1.25em;
-    font-weight: bold;
-    text-align: center;
-    color: ${props => props.theme.skyBlue};
-    @media ${DESKTOP} {
-      font-size: ${props => (props.size === 'small' ? '1.25em' : '1.5em')};
-    }
-  }
-
-  ${FooterStyled} {
-    border-top: solid 1px ${props => hexToRgba(props.theme.stoneGrey, 0.5)};
-    padding: ${props => PADDINGS[props.size] || PADDINGS.normal};
-    display: flex;
-    flex-flow: column nowrap;
-    align-items: center;
-
-    > * + * {
-      margin-top: 20px;
-    }
-
-    ${props =>
-    props.size !== 'small'
-      && `
-      @media ${DESKTOP} {
-        flex-flow: row-reverse wrap;
-        align-items: center;
-        justify-content: space-between;
-        > * + * {
-          margin-top: 0;
-        }
-      }
-    `};
-  }
-
-  ${BodyStyled} {
-    padding: ${props => PADDINGS[props.size] || PADDINGS.normal};
-  }
-`;
-
-const Root = styled.section`
-  .ft-modal--backdrop {
-    position: fixed;
-    display: block;
-    top: 0;
-    right: 0;
-    bottom: 0;
-    left: 0;
-    max-height: 100vh;
-    background-color: ${props => props.backdropColor || 'rgba(0, 0, 0, 0.61)'};
-    z-index: 9999;
-    overflow-x: hidden;
-    overflow-y: auto;
-  }
-
-  .ft--modal--dialog {
-    z-index: 10000;
-    position: absolute;
-    max-height: 100vh;
-    &:focus {
-      outline: none;
-    }
-    @media ${DESKTOP} {
-      top: 50%;
-      left: 50%;
-      transform: translate(-50%, -50%);
-    }
-  }
+const Backdrop = styled.section`
+  position: fixed;
+  display: block;
+  top: 0;
+  right: 0;
+  bottom: 0;
+  left: 0;
+  max-height: 100vh;
+  background-color: ${props => props.backdropColor || 'rgba(0, 0, 0, 0.61)'};
+  z-index: 9999;
+  overflow-x: hidden;
+  overflow-y: auto;
 `;
 
 const ModalWrapper = styled.div`
-  ${ModalStyles};
+  z-index: 10000;
+  position: absolute;
+  max-height: 100vh;
+  width: 100%;
+  top: 50%;
+  left: 50%;
+  transform: translate(-50%, -50%);
+  &:focus {
+    outline: none;
+  }
+  @media ${DESKTOP} {
+    width: auto;
+  }
+`;
+
+const Content = styled.div`
   height: 100vh;
   @media ${DESKTOP} {
     height: auto;
@@ -120,66 +44,106 @@ const ModalWrapper = styled.div`
 `;
 
 /**
- * Main modal component.
+ * Modal component.
  */
 class Modal extends Component {
-  preventBodyOverflow = null;
+  previousBodyOverflow = null;
 
   componentDidMount() {
-    const { open, preventGlobalScroll } = this.props;
-    if (open && preventGlobalScroll) {
-      this.blockGlobalScroll();
+    const { open } = this.props;
+    if (open) {
+      this.handleShow();
     }
   }
 
   componentDidUpdate({ open: prevOpen }) {
-    const { open, preventGlobalScroll } = this.props;
-    if (!preventGlobalScroll) {
-      return;
-    }
+    const { open } = this.props;
     if (!prevOpen && open) {
-      this.blockGlobalScroll();
+      this.handleShow();
     } else if (prevOpen && !open) {
+      this.handleHide();
+    }
+  }
+
+  componentWillUnmount() {
+    this.handleHide();
+  }
+
+  addEventListeners = () => {
+    if (!document) return;
+    document.addEventListener('keydown', this.handleDocumentKeydown);
+  };
+
+  removeEventListeners = () => {
+    if (!document) return;
+    document.removeEventListener('keydown', this.handleDocumentKeydown);
+  };
+
+  blockGlobalScroll = () => {
+    if (!document) return;
+    this.previousBodyOverflow = document.body.style.overflow;
+    document.body.style.overflow = 'hidden';
+  };
+
+  unblockGlobalScroll = () => {
+    if (!document) return;
+    document.body.style.overflow = this.previousBodyOverflow;
+    this.previousBodyOverflow = null;
+  };
+
+  handleShow = () => {
+    focusStore.storeFocus();
+    focusScope.scopeFocus(this.modal);
+    this.addEventListeners();
+    if (this.props.preventGlobalScroll) {
+      this.blockGlobalScroll();
+    }
+  };
+
+  handleHide = () => {
+    focusScope.unscopeFocus();
+    focusStore.restoreFocus();
+    this.removeEventListeners();
+    if (this.props.preventGlobalScroll) {
       this.unblockGlobalScroll();
     }
-  }
+  };
 
-  blockGlobalScroll() {
-    if (!document) return;
-    this.preventBodyOverflow = document.body.style.overflow;
-    document.body.style.overflow = 'hidden';
-  }
+  handleDocumentKeydown = event => {
+    if (event.keyCode === 27) {
+      this.props.onClose();
+    }
+  };
 
-  unblockGlobalScroll() {
-    if (!document) return;
-    document.body.style.overflow = this.preventBodyOverflow;
-    this.preventBodyOverflow = null;
-  }
+  handleBackdropClick = () => {
+    this.props.onClose();
+  };
+
+  handleModalClick = event => {
+    event.stopPropagation();
+  };
+
+  handleRef = element => {
+    this.modal = element;
+  };
 
   render() {
-    const { backdropColor, open, size, children, onClose } = this.props;
+    const { backdropColor, open, children } = this.props;
     if (!open) {
       return null;
     }
     return (
-      <Root backdropColor={backdropColor} size={size} role="dialog">
-        <BaseModal
-          onClose={onClose}
-          closeOnEsc
-          closeOnBackdropClick
-          backdropClassName="ft-modal--backdrop"
-          modalClassName="ft--modal--dialog"
-        >
-          <ModalWrapper size={size}>{children}</ModalWrapper>
-        </BaseModal>
-      </Root>
+      <Backdrop backdropColor={backdropColor} role="dialog" onClick={this.handleBackdropClick}>
+        <ModalWrapper innerRef={this.handleRef} onClick={this.handleModalClick} tabIndex="-1">
+          <Content>{children}</Content>
+        </ModalWrapper>
+      </Backdrop>
     );
   }
 }
 
 Modal.defaultProps = {
   preventGlobalScroll: true,
-  size: 'normal',
 };
 
 Modal.propTypes = {
@@ -187,16 +151,8 @@ Modal.propTypes = {
   children: PropTypes.node.isRequired,
   preventGlobalScroll: PropTypes.bool,
   open: PropTypes.bool.isRequired,
-  size: PropTypes.oneOf(['auto', 'small', 'normal', 'large']),
   onClose: PropTypes.func,
 };
 
-Modal.Header = HeaderStyled;
-Modal.Footer = FooterStyled;
-Modal.Body = BodyStyled;
-
 /** @component */
 export default Modal;
-
-// Screen-readers support opt-out. See https://github.com/cloudflare/react-modal2#accessibility
-BaseModal.getApplicationElement = () => null;