+++ CHANGELOG.md
@@ -1,4 +1,9 @@
 ## [Unreleased]
+
+### Changed
+- Split Modal component to Modal and Dialog components.
+- Get rid of `react-modal2` dependency.
+
 ### Added
 - Checkbox component